requirements.txt
numpy
opencv-python
pickle-mixin
requests
torch
scikit-image
pillow